+++ src/Storages/MergeTree/MergeTreeReadPool.h
@@ -1,10 +1,13 @@
 #pragma once
 
 #include <Core/NamesAndTypes.h>
-#include <Storages/MergeTree/RangesInDataPart.h>
+#include <Storages/MergeTree/MergeTreeBaseSelectProcessor.h>
 #include <Storages/MergeTree/MergeTreeBlockReadUtils.h>
 #include <Storages/MergeTree/MergeTreeData.h>
+#include <Storages/MergeTree/RangesInDataPart.h>
+#include <Storages/MergeTree/RequestResponse.h>
 #include <Storages/SelectQueryInfo.h>
+
 #include <mutex>
 
 
@@ -17,10 +20,56 @@ using MergeTreeReadTaskPtr = std::unique_ptr<MergeTreeReadTask>;
 class IMergeTreeReadPool
 {
 public:
+    IMergeTreeReadPool(
+        StorageSnapshotPtr storage_snapshot_,
+        Names column_names_,
+        Names virtual_column_names_,
+        size_t min_marks_for_concurrent_read_,
+        PrewhereInfoPtr prewhere_info_,
+        RangesInDataParts parts_ranges_,
+        bool predict_block_size_bytes_,
+        bool do_not_steal_tasks_)
+        : storage_snapshot(storage_snapshot_)
+        , column_names(column_names_)
+        , virtual_column_names(virtual_column_names_)
+        , min_marks_for_concurrent_read(min_marks_for_concurrent_read_)
+        , prewhere_info(prewhere_info_)
+        , parts_ranges(parts_ranges_)
+        , predict_block_size_bytes(predict_block_size_bytes_)
+        , do_not_steal_tasks(do_not_steal_tasks_)
+    {}
+
     virtual MergeTreeReadTaskPtr getTask(size_t thread) = 0;
     virtual Block getHeader() const = 0;
     virtual void profileFeedback(ReadBufferFromFileBase::ProfileInfo info) = 0;
     virtual ~IMergeTreeReadPool() = default;
+
+protected:
+
+    std::vector<size_t> fillPerPartInfo(const RangesInDataParts & parts);
+
+    /// Initialized in constructor
+    StorageSnapshotPtr storage_snapshot;
+    const Names column_names;
+    const Names virtual_column_names;
+    size_t min_marks_for_concurrent_read{0};
+    PrewhereInfoPtr prewhere_info;
+    RangesInDataParts parts_ranges;
+    bool predict_block_size_bytes;
+    bool do_not_steal_tasks;
+
+    struct PerPartParams
+    {
+        MergeTreeReadTaskColumns task_columns;
+        NameSet column_name_set;
+        MergeTreeBlockSizePredictorPtr size_predictor;
+        RangesInDataPart data_part;
+    };
+
+    std::vector<PerPartParams> per_part_params;
+    std::vector<bool> is_part_on_remote_disk;
+
+    mutable std::mutex mutex;
 };
 
 using IMergeTreeReadPoolPtr = std::shared_ptr<IMergeTreeReadPool>;
@@ -106,37 +155,10 @@ public:
     Block getHeader() const override;
 
 private:
-    std::vector<size_t> fillPerPartInfo(const RangesInDataParts & parts);
 
     void fillPerThreadInfo(
         size_t threads, size_t sum_marks, std::vector<size_t> per_part_sum_marks,
         const RangesInDataParts & parts);
-
-    StorageSnapshotPtr storage_snapshot;
-    const Names column_names;
-    const Names virtual_column_names;
-    bool do_not_steal_tasks;
-    bool predict_block_size_bytes;
-    size_t min_marks_for_concurrent_read{0};
-
-    struct PerPartParams
-    {
-        MergeTreeReadTaskColumns task_columns;
-        NameSet column_name_set;
-        MergeTreeBlockSizePredictorPtr size_predictor;
-    };
-
-    std::vector<PerPartParams> per_part_params;
-
-    PrewhereInfoPtr prewhere_info;
-
-    struct Part
-    {
-        MergeTreeData::DataPartPtr data_part;
-        size_t part_index_in_query;
-    };
-
-    std::vector<Part> parts_with_idx;
 
     struct ThreadTask
     {
@@ -151,18 +173,104 @@ private:
     };
 
     std::vector<ThreadTask> threads_tasks;
-
     std::set<size_t> remaining_thread_tasks;
-
-    RangesInDataParts parts_ranges;
-
-    mutable std::mutex mutex;
-
     Poco::Logger * log = &Poco::Logger::get("MergeTreeReadPool");
 
-    std::vector<bool> is_part_on_remote_disk;
 };
 
 using MergeTreeReadPoolPtr = std::shared_ptr<MergeTreeReadPool>;
+
+class MergeTreeReadPoolParallelReplicas : public IMergeTreeReadPool, private boost::noncopyable
+{
+public:
+
+    MergeTreeReadPoolParallelReplicas(
+        StorageSnapshotPtr storage_snapshot_,
+        size_t threads_,
+        ParallelReadingExtension extension_,
+        const RangesInDataParts & parts_,
+        const PrewhereInfoPtr & prewhere_info_,
+        const Names & column_names_,
+        const Names & virtual_column_names_,
+        size_t min_marks_for_concurrent_read_
+    )
+    : IMergeTreeReadPool(
+        storage_snapshot_,
+        column_names_,
+        virtual_column_names_,
+        min_marks_for_concurrent_read_,
+        prewhere_info_,
+        parts_,
+        /*predict_block_size*/false,
+        /*do_not_steal_tasks*/false)
+    , extension(extension_)
+    , threads(threads_)
+    {
+        fillPerPartInfo(parts_ranges);
+
+        extension.all_callback({
+            .description = parts_ranges.getDescriptions(),
+            .replica_num = extension.number_of_current_replica
+        });
+    }
+
+    ~MergeTreeReadPoolParallelReplicas() override;
+
+    MergeTreeReadTaskPtr getTask(size_t thread) override;
+    Block getHeader() const override;
+    void profileFeedback(ReadBufferFromFileBase::ProfileInfo) override {}
+
+private:
+    ParallelReadingExtension extension;
+
+    RangesInDataPartsDescription buffered_ranges;
+    size_t threads;
+    bool no_more_tasks_available{false};
+    Poco::Logger * log = &Poco::Logger::get("MergeTreeReadPoolParallelReplicas");
+};
+
+using MergeTreeReadPoolParallelReplicasPtr = std::shared_ptr<MergeTreeReadPoolParallelReplicas>;
+
+
+class MergeTreeInOrderReadPoolParallelReplicas : private boost::noncopyable
+{
+public:
+    MergeTreeInOrderReadPoolParallelReplicas(
+        RangesInDataParts parts_,
+        ParallelReadingExtension extension_,
+        CoordinationMode mode_,
+        size_t min_marks_for_concurrent_read_)
+    : parts_ranges(parts_)
+    , extension(extension_)
+    , mode(mode_)
+    , min_marks_for_concurrent_read(min_marks_for_concurrent_read_)
+    {
+        for (const auto & part : parts_ranges)
+            request.push_back({part.data_part->info, MarkRanges{}});
+
+        for (const auto & part : parts_ranges)
+            buffered_tasks.push_back({part.data_part->info, MarkRanges{}});
+
+        extension.all_callback({
+            .description = parts_ranges.getDescriptions(),
+            .replica_num = extension.number_of_current_replica
+        });
+    }
+
+    MarkRanges getNewTask(RangesInDataPartDescription description);
+
+    RangesInDataParts parts_ranges;
+    ParallelReadingExtension extension;
+    CoordinationMode mode;
+    size_t min_marks_for_concurrent_read{0};
+
+    bool no_more_tasks{false};
+    RangesInDataPartsDescription request;
+    RangesInDataPartsDescription buffered_tasks;
+
+    std::mutex mutex;
+};
+
+using MergeTreeInOrderReadPoolParallelReplicasPtr = std::shared_ptr<MergeTreeInOrderReadPoolParallelReplicas>;
 
 }
